+++ src/lib/components/Card/Card.module.scss
