+++ src/lib/components/CardSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardSkeleton';
