+++ src/lib/components/ReviewCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ReviewCard';
